requirements.txt
streamlit
mediapipe
opencv-python
numpy
SpeechRecognition
pillow